deck: extract shared DeckMenu from player and compact bar
